[Fix] Restore fallback quad context and shader bindings

+++ src/main/java/com/dhj/actinium/celeritas/vertices/BlockRenderContext.java
@@ -4,11 +4,11 @@ public class BlockRenderContext {
     public int localPosX;
     public int localPosY;
     public int localPosZ;
-    public short blockId = -1;
+    public int blockId = -1;
     public short renderType = -1;
     public byte lightValue;
 
-    public void set(int localX, int localY, int localZ, short blockId, short renderType, byte lightValue) {
+    public void set(int localX, int localY, int localZ, int blockId, short renderType, byte lightValue) {
         this.localPosX = localX;
         this.localPosY = localY;
         this.localPosZ = localZ;
